requirements.txt
python-dotenv
neofuzz
embetter
openai
gliner
fastapi
uvicorn
pydantic
